publisher/requirements.txt
aiogram==2.25.1
aiogram-dialog==1.9.0
aiohttp==3.8.4 ; python_version >= '3.6'
aiosignal==1.3.1 ; python_version >= '3.7'
async-timeout==4.0.2 ; python_version >= '3.6'
attrs==22.2.0 ; python_version >= '3.6'
babel==2.9.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
black==23.1.0
cachetools==4.2.4 ; python_version ~= '3.5'
certifi==2022.12.7 ; python_version >= '3.6'
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
click==8.1.3 ; python_version >= '3.7'
frozenlist==1.3.3 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
jinja2==3.1.2 ; python_version >= '3.7'
magic-filter==1.0.9 ; python_version >= '3.7' and python_version < '4.0'
markupsafe==2.1.2 ; python_version >= '3.7'
multidict==6.0.4 ; python_version >= '3.7'
mypy-extensions==1.0.0 ; python_version >= '3.5'
packaging==23.0 ; python_version >= '3.7'
pathspec==0.11.1 ; python_version >= '3.7'
platformdirs==3.1.1 ; python_version >= '3.7'
pydantic==1.9.1
pytz==2022.7.1
pyyaml==6.0
requests==2.28.2
tomli==2.0.1 ; python_version < '3.11'
typing-extensions==4.5.0 ; python_version >= '3.7'
urllib3==1.26.15 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
yarl==1.8.2 ; python_version >= '3.7'